pyproject.toml
[project]
name = "graphiti-poc"
version = "0.1.0"
description = "远程Graphiti知识图谱客户端"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    # 核心依赖
    "graphiti-core",
    "neo4j>=5.14.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "openai>=1.0.0",
    "aiohttp>=3.8.0",
    "pydantic>=2.0.0",
    "tenacity>=8.0.0",
    "typing-extensions>=4.0.0",
    # Web框架依赖
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
]

[project.optional-dependencies]
# 开发依赖
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

# 脚本入口（可选）
[project.scripts]
graphiti-demo = "main:main"
graphiti-api = "api.main:start_server"

#[build-system]
#requires = ["hatchling"]
#build-backend = "hatchling.build"

[tool.uv]
# uv 特有配置
cache-dir = ".uv-cache"